refactor: dayjs and camel case
